python-training/requirements.txt
torch
transformers
trl
peft
bitsandbytes
datasets
google-cloud-storage
accelerate